// src/lib/config/map.types.ts
type themes = {
	one: baseValues;
	two: baseValues;
	three: baseValues;
	four: baseValues;
	five?: baseValues;
	six?: baseValues;
	seven?: baseValues;
	eight?: baseValues;
	nine?: baseValues;
	ten?: baseValues;
};

type baseValues = {
	one: string;
	two: string;
	three?: string;
	four?: string;
	five?: string;
	six?: string;
	seven?: string;
	eight?: string;
	nine?: string;
	ten?: string;
	eleven?: string;
	twelve?: string;
};

export type { themes, baseValues };